modified install to accept commandline arguments

+++ install.sh
@@ -1,5 +1,8 @@
 #!/bin/bash
 
+# Installs all config files in this repo. Specify for which tools to install
+# with command line arguments.
+
 # Define folder and target location
 declare -A folders
 folders[alacritty]="$HOME/.config/alacritty"
@@ -7,33 +10,36 @@ folders[fish]="$HOME/.config/fish"
 folders[i3]="$HOME/.config/i3"
 folders[nvim]="$HOME/.config/nvim"
 
-# Loop and symlink folders
-for folder in "${!folders[@]}"; do
-    target="${folders[$folder]}"
-    source="$PWD/$folder"
+declare -A files
+files[tmux]="$HOME/.tmux.conf"
+files[zsh]="$HOME/.zshrc"
+
+# Loop through the command-line arguments
+for tool in "$@"; do
+  # Symlink folders
+  if [ "${folders[$tool]+isset}" ]; then
+    target="${folders[$tool]}"
+    source="$PWD/$tool"
 
     if [ -d "$target" ] || [ -L "$target" ]; then
-        rm -rf "$target"
+      rm -rf "$target"
     fi
 
     ln -s "$source" "$target"
     echo "Symlinked $source to $target"
-done
+  fi
 
-# Handle individual files
-declare -A files
-files[tmux/.tmux.conf]="$HOME/.tmux.conf"
-files[zsh/.zshrc]="$HOME/.zshrc"
-
-for file in "${!files[@]}"; do
-    target="${files[$file]}"
-    source="$PWD/$file"
+  # Symlink individual files
+  if [ "${files[$tool]+isset}" ]; then
+    target="${files[$tool]}"
+    source="$PWD/$tool/.tmux.conf"
 
     if [ -f "$target" ] || [ -L "$target" ]; then
-        rm -rf "$target"
+      rm -rf "$target"
     fi
 
     ln -s "$source" "$target"
     echo "Symlinked $source to $target"
+  fi
 done
